added dashboard listing UI

+++ lib/utils/size_utils.dart
@@ -8,7 +8,7 @@ extension SizeExtension on BuildContext {
   double get width => size.width;
   double get height => size.height;
 
-  double widthPercent(int percent) => width * percent/100;
-  double heightPercent(int percent) => height * percent/100;
+  double widthPercent(double percent) => width * percent/100;
+  double heightPercent(double percent) => height * percent/100;
 
 }
